+++ src/greeklib/xstrings.proto.h
@@ -4,4 +4,4 @@ int Xstrcpy(char *, const char *);
 int Xstrncpy(char *, const char *, size_t);
 void Xstrncat(char *, const char *, size_t);
 int Xstrncmp(const char *, const char *, size_t);
-int Xstrlen(const char *);
+size_t Xstrlen(const char *);
