+++ src/morphlib/beta2rtf.proto.h
@@ -1,7 +1,9 @@
 
 /* beta2rtf.c */
+#include <stddef.h>
+
 int _main(int, char *[]);
 void conv_defline(char *, FILE *);
-int check_deflev(char *, char *, int);
+int check_deflev(char *, char *, size_t);
 int has_pref(char *, char *);
 int is_greek(char *);
